Move commands into e2e subdir
