+++ cypress/support/commands/e2e/auth.js
