+++ cypress/support/commands/e2e/index.js
@@ -14,4 +14,5 @@
 // ***********************************************************
 
 import '@testing-library/cypress/add-commands';
-import './commands';
+import './api';
+import './auth';
